+++ cardano-api/gen/Test/Gen/Cardano/Api/Typed.hs
@@ -125,7 +125,7 @@ import           Cardano.Api.Shelley (GovernancePoll (..), GovernancePollAnswer 
                    PlutusScript (PlutusScriptSerialised), ProtocolParameters (ProtocolParameters),
                    ReferenceScript (..), ReferenceTxInsScriptsInlineDatumsSupportedInEra (..),
                    StakeCredential (StakeCredentialByKey), StakePoolKey,
-                   refInsScriptsAndInlineDatsSupportedInEra)
+                   refInsScriptsAndInlineDatsSupportedInEra, scriptInEraToRefScript)
 
 import qualified Cardano.Binary as CBOR
 import qualified Cardano.Crypto.Hash as Crypto
@@ -512,7 +512,7 @@ genReferenceScript :: CardanoEra era -> Gen (ReferenceScript era)
 genReferenceScript era =
   case refInsScriptsAndInlineDatsSupportedInEra era of
     Nothing -> return ReferenceScriptNone
-    Just supp -> ReferenceScript supp <$> genScriptInAnyLang
+    Just _ -> scriptInEraToRefScript <$> genScriptInEra era
 
 genUTxO :: CardanoEra era -> Gen (UTxO era)
 genUTxO era =