+++ cardano-api/ChangeLog.md
@@ -3,6 +3,7 @@
 ## New
 
 - **Breaking change** Remove support for reading `ApplicationName` and `ApplicationVersion` from node configuration and replace with hardcoded values. ([PR 8](https://github.com/input-output-hk/cardano-api/pull/8))
+- Updated ledger dependency bounds, added support for Plutus V3 in conway
 
 ## 8.1.0